dex(router): move tests/clean deps
