+++ app/src/dex/router/fill_route.rs
@@ -1,5 +1,3 @@
-use std::collections::BTreeMap;
-
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use penumbra_crypto::{
